Untested support for 5-atom algebras
I've added a gen5Atoms function that will, in theory, be able to generate all 5-atom NAs in 9 different parts (to reduce the isomorphism checking load). However, I may need a bit more computational resources to run this function.

I'm unhappy with the repetitive way in which the gen4Atoms and gen5Atoms functions are written. I think I'll create a wrapping function that can clear these up a bit.