Added the option to export algebras to text
The generate_algebras function now supports an export Boolean argument, defaulted to False, which, if enabled, will export algebras as 2-tuples. The first element of the 2-tuple is the atom table of the algebra and the second element is the converse (as pairs). This is useful if we wish to export the algebras to a text file, where they can alter be created as AtomicAlgebras.